refactor(mediawiki): extract list and table conversion from converter.go (deslop #59)
