+++ converter/converter.go
@@ -237,169 +237,6 @@ func detectLanguage(code string) string {
 	return "text"
 }
 
-// listItem tracks list type at each indent level
-type listItem struct {
-	indentLevel int
-	listType    string // "*" or "#"
-}
-
-// convertLists converts Markdown lists to MediaWiki format
-func convertLists(text string) string {
-	lines := strings.Split(text, "\n")
-	result := make([]string, 0, len(lines))
-
-	unorderedRegex := regexp.MustCompile(`^(\s*)[-\*]\s+(.*)$`)
-	orderedRegex := regexp.MustCompile(`^(\s*)\d+\.\s+(.*)$`)
-
-	var listStack []listItem
-
-	for _, line := range lines {
-		if matches := unorderedRegex.FindStringSubmatch(line); matches != nil {
-			indent := len(matches[1])
-			content := matches[2]
-			currentLevel := indent / 2
-
-			prefix := buildListPrefix(listStack, currentLevel, "*")
-			line = prefix + " " + content
-			listStack = updateListStack(listStack, currentLevel, "*")
-
-		} else if matches := orderedRegex.FindStringSubmatch(line); matches != nil {
-			indent := len(matches[1])
-			content := matches[2]
-			currentLevel := indent / 2
-
-			prefix := buildListPrefix(listStack, currentLevel, "#")
-			line = prefix + " " + content
-			listStack = updateListStack(listStack, currentLevel, "#")
-
-		} else {
-			listStack = nil
-		}
-
-		result = append(result, line)
-	}
-
-	return strings.Join(result, "\n")
-}
-
-func buildListPrefix(stack []listItem, currentLevel int, currentType string) string {
-	prefix := ""
-	for i := 0; i <= currentLevel && i < len(stack); i++ {
-		prefix += stack[i].listType
-	}
-	if currentLevel >= len(stack) {
-		prefix += currentType
-	}
-	return prefix
-}
-
-func updateListStack(stack []listItem, currentLevel int, currentType string) []listItem {
-	if currentLevel < len(stack) {
-		stack = stack[:currentLevel]
-	}
-	if currentLevel < len(stack) {
-		stack[currentLevel].listType = currentType
-	} else {
-		stack = append(stack, listItem{
-			indentLevel: currentLevel,
-			listType:    currentType,
-		})
-	}
-	return stack
-}
-
-// convertTables converts Markdown tables to MediaWiki format
-// splitTableCells extracts the middle cells from a pipe-bounded line.
-func splitTableCells(line string) []string {
-	cells := strings.Split(line, "|")
-	if len(cells) <= 2 {
-		return nil
-	}
-	return cells[1 : len(cells)-1]
-}
-
-// cellsAreAllDashes reports whether every cell is a separator (only - and :).
-func cellsAreAllDashes(cells []string) bool {
-	for _, cell := range cells {
-		if strings.TrimSpace(strings.ReplaceAll(strings.ReplaceAll(cell, "-", ""), ":", "")) != "" {
-			return false
-		}
-	}
-	return true
-}
-
-// emitTableHeader appends a wikitable opening + header cells from the markdown
-// header line and returns the index just past any separator row.
-func emitTableHeader(line string, lines []string, i int, result *[]string) int {
-	*result = append(*result, `{| class="wikitable"`)
-	cells := strings.Split(line, "|")
-	cells = cells[1 : len(cells)-1]
-	*result = append(*result, "|-")
-	for _, cell := range cells {
-		*result = append(*result, "! "+strings.TrimSpace(cell))
-	}
-	i++
-	if i < len(lines) {
-		nextLine := strings.TrimSpace(lines[i])
-		if strings.Contains(nextLine, "|") && strings.Contains(nextLine, "-") {
-			i++
-		}
-	}
-	return i
-}
-
-// emitTableBodyRow appends a non-header row's cells, skipping pure separator rows.
-func emitTableBodyRow(line string, result *[]string) {
-	cells := splitTableCells(line)
-	if cells == nil || cellsAreAllDashes(cells) {
-		return
-	}
-	*result = append(*result, "|-")
-	for _, cell := range cells {
-		*result = append(*result, "| "+strings.TrimSpace(cell))
-	}
-}
-
-func convertTables(text string) string {
-	lines := strings.Split(text, "\n")
-	result := make([]string, 0, len(lines))
-	inTable := false
-
-	pipeRegex := regexp.MustCompile(`^\|.*\|$`)
-	separatorRegex := regexp.MustCompile(`^\|[\s\-:|]+\|$`)
-
-	i := 0
-	for i < len(lines) {
-		line := strings.TrimSpace(lines[i])
-
-		if !inTable && strings.Contains(line, "|") && pipeRegex.MatchString(line) {
-			inTable = true
-			i = emitTableHeader(line, lines, i, &result)
-			continue
-		}
-		if inTable && strings.Contains(line, "|") {
-			if !(separatorRegex.MatchString(line) && strings.Contains(line, "-")) {
-				emitTableBodyRow(line, &result)
-			}
-			i++
-			continue
-		}
-		if inTable && !strings.Contains(line, "|") {
-			result = append(result, "|}")
-			inTable = false
-		}
-
-		result = append(result, lines[i])
-		i++
-	}
-
-	if inTable {
-		result = append(result, "|}")
-	}
-
-	return strings.Join(result, "\n")
-}
-
 // convertHorizontalRules converts Markdown hr to MediaWiki
 func convertHorizontalRules(text string) string {
 	hrRegex := regexp.MustCompile(`(?m)^[\s]*[-*_]{3,}[\s]*$`)
